Move example functions into separate folder and clean up index.ts
